Refactor production order helpers
